Set current working directory of ocamlformat child process
Ocamlformat relies on the location of the file being formatted in order
to search for a 'project root' which may contain an .ocamlformat
configuration file. Since the invocation of ocamlformat within the
vscode extension uses stdin instead of a named file, ocamlformat does
not know the location of the file and therefore relies on the current
working directory of the ocamlformat invocation to search for the
project root instead. The cwd of the ocamlformat child process is
undefined and as such can sometimes work and sometimes fail with
'project root not found' errors. This commit ensures that the cwd of the
ocamlformat process is set to the directory of the file which is being
formatted, to ensure that the correct project root is found and
therefore the appropriate .ocamlformat configuration is used each time.